+++ packages/@dicebear/croodles-neutral/src/components/index.ts
@@ -1,0 +1,3 @@
+export { eyes } from './eyes';
+export { nose } from './nose';
+export { mouth } from './mouth';