+++ packages/@dicebear/croodles/src/components/index.ts
@@ -1,0 +1,7 @@
+export { face } from './face';
+export { nose } from './nose';
+export { beard } from './beard';
+export { mouth } from './mouth';
+export { top } from './top';
+export { mustache } from './mustache';
+export { eyes } from './eyes';