+++ website/src/options/croodlesNeutral.ts
@@ -1,0 +1,5 @@
+import defaults from './defaults';
+
+export default {
+  ...defaults,
+};